# Conflicts:
#	app/src/main/java/ru/smartro/worknote/presentation/PMapF.kt
#	app/src/main/java/ru/smartro/worknote/presentation/checklist/workorder/StartWorkOrderViewModel.kt
#	app/src/main/java/ru/smartro/worknote/work/WorkOrderResponse_know1.kt

+++ app/src/main/java/ru/smartro/worknote/awORKOLDs/SynchroOidWidRequestPOST.kt
@@ -156,7 +156,7 @@ data class SynchroOidWidOutBodyDataWorkorderPlatform(
     var updateAt: Long,
     @Expose
     @SerializedName("srp_id")
-    val srpId: Int,
+    val srpId: Long,
     @Expose
     @SerializedName("order_start_time")
     var orderStartTime: String? = null,
